+++ R/globals.R
@@ -1,0 +1,1 @@
+.globals <- new.env(hash = TRUE)